refactor: 컴포넌트명과 파일명이 일치하도록 수정
