+++ frontend/src/pages/recommend/components/ReadingKingLeaderboard/LeaderboardItem.tsx
